refactor: move validation into separate file, fix types a bit, create separate comp for an error
