+++ src/components/inputs/TextInput.tsx
@@ -1,13 +1,13 @@
 "use client";
 
-import { FieldErrors, FieldValues, UseFormRegister } from "react-hook-form";
-import { QuestionMarkCircleIcon } from "@heroicons/react/20/solid";
+import { FieldErrors, UseFormRegisterReturn } from "react-hook-form";
+import ErrorMessage from "./ErrorMessage";
 
 interface TextInputProps {
   id: string;
   label?: string;
   placeholder: string;
-  register: UseFormRegister<FieldValues>;
+  register: UseFormRegisterReturn;
   errors: FieldErrors;
 }
 
@@ -31,21 +31,14 @@ export default function TextInput({
         id={id}
         type="text"
         placeholder={placeholder}
-        {...register(id)}
+        {...register}
         className={`w-full my-2 text-base shadow-transparent shadow-[0px_0px_0px_3px] input input-bordered bg-base-200 text-neutral-focus ${
           errors[id]
             ? "border-error focus-visible:shadow-error/30"
             : "border-neutral/40 focus-visible:shadow-neutral/30"
         }`}
       />
-      <label className="p-0 pt-1 label">
-        <span className="flex items-center h-4 gap-1 text-sm label-text-alt text-error">
-          {errors[id] && (
-            <QuestionMarkCircleIcon className="w-4 h-4 stroke-error fill-transparent" />
-          )}
-          {errors[id]?.message as string}
-        </span>
-      </label>
+      <ErrorMessage message={errors[id]?.message as string} />
     </div>
   );
 }
