+++ src/components/modals/index.tsx
@@ -1,2 +1,2 @@
-export { default as Example1Modal } from "./Example1Modal";
+export { default as Example1Modal } from "./Example1Modal/Example1Modal";
 export { default as Example2Modal } from "./Example2Modal";
